SwotAnalysis/mcp_server_demo/requirements.txt
flask==2.3.2
requests==2.31.0
pandas==1.5.3
matplotlib==3.7.2
transformers==4.30.2
fpdf==1.7.2
torch==2.0.1
gunicorn==21.2.0
python-dotenv